ab-deliveries-starter/server-python/requirements.txt
fastapi
uvicorn[standard]
motor
passlib[bcrypt]
pydantic[email]
python-dotenv